xnor-sdk/samples/rpi3/python/requirements.txt
Pillow>=5.1.0
picamera
psutil
